Single stage Docker build
Build the Docker image in a single stage instead of using a multi-stage
layout. The build stage is replaced with a script that's run separately
to produce the required binary. This allows caching the dependencies on
each build.

This commit also changes the final Docker image from Alpine to
Distroless.